LPD-67164 move testControlChannelProperties to a new ClusterLinkTest since it covers cluster.link channel configuration rather than general cluster behavior
